Moved every file to its own places + finished the onboarding sites - Missing animations
