+++ frontend/src/components/css/profileCompletionOne.css
@@ -4,10 +4,6 @@
 #ProfileCompletion-Wrapper {
   height: 100%;
   display: flex;
-}
-#ProfileCompletion-wrapper p,
-h2,
-h3 {
   font-family: "Montserrat", sans-serif;
 }
 .Profile-completion-one-right {
@@ -24,12 +20,55 @@ h3 {
   width: 50%;
   min-width: 50%;
 }
+.Profile-completion-one-content .Description p,
+.User-photos p {
+  color: #000;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
 .Profile-completion-one-right .Header-text {
   color: #000;
   font-size: 30px;
   font-style: normal;
   font-weight: 800;
   line-height: normal;
+}
+.Profile-completion-one-content .User-photos {
+  position: relative;
+  margin-bottom: 5%;
+}
+.Profile-completion-one-content .User-photos button {
+  width: 100%;
+  height: 156px;
+  border-radius: 8px;
+  background: rgba(217, 217, 217, 0.38);
+}
+.Profile-completion-one-content #upload-image {
+  display: none;
+}
+.Profile-completion-one-content .Circle {
+  width: 100px;
+  height: 100px;
+  border: 1px solid grey;
+  display: flex;
+  border-radius: 50%;
+  background: rgb(217, 217, 217);
+  position: absolute;
+  top: 72.5%;
+  left: 2.5%;
+}
+.Profile-completion-one-content .Circle.centered {
+  justify-content: center;
+  align-items: center;
+}
+
+.Profile-completion-one-content .Circle img {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  object-fit: cover;
 }
 .Profile-completion-one-content .Description {
   max-width: 100%;
